refactor: ComputerNumber, UserNumber 클래스 위치 변경
기존 domain 패키지에서 controller 패키지로 이동
generate 기능만 수행하는 위 클래스들은 controller 레이어에 포함되어야 하기 때문에 수정
